# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Menu Engineering API",
        "description": "API para gestão de engenharia de cardápio",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/batches": {
            "get": {
                "tags": [
                    "Batches"
                ],
                "operationId": "78f7a2aaa0170b07860cd40cccc9e601",
                "responses": {
                    "200": {
                        "description": "Lista"
                    }
                }
            },
            "post": {
                "tags": [
                    "Batches"
                ],
                "operationId": "500a3c8e012b687d0030233a4a527ae6",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Criado"
                    }
                }
            }
        },
        "/api/batches/{id}": {
            "get": {
                "tags": [
                    "Batches"
                ],
                "operationId": "717f647d6a2044806c2279b73614b720",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhe"
                    }
                }
            },
            "put": {
                "tags": [
                    "Batches"
                ],
                "operationId": "49c286f5608e8f6c08e496bf0611f196",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Atualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Batches"
                ],
                "operationId": "efe66d7ff74927d2e8b0650babbbf400",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removido"
                    }
                }
            }
        },
        "/api/dishes": {
            "get": {
                "tags": [
                    "Dishes"
                ],
                "operationId": "ddf841d9ba35d917bd22a504c13b2f72",
                "responses": {
                    "200": {
                        "description": "Lista"
                    }
                }
            },
            "post": {
                "tags": [
                    "Dishes"
                ],
                "operationId": "3fc0252a7de74dac3747d26739c844cb",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Criado"
                    }
                }
            }
        },
        "/api/dishes/{id}": {
            "get": {
                "tags": [
                    "Dishes"
                ],
                "operationId": "07b6fe7c8dc79b35e66b58956b97b893",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhe"
                    }
                }
            },
            "put": {
                "tags": [
                    "Dishes"
                ],
                "operationId": "52b7975eeaea935e1251d215f5ac6bb4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Atualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Dishes"
                ],
                "operationId": "57bb2309cba8e67241d5d84d8ceb6beb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removido"
                    }
                }
            }
        },
        "/api/ingredients": {
            "get": {
                "tags": [
                    "Ingredients"
                ],
                "operationId": "88554df920fe97e07f179ca559f98b75",
                "responses": {
                    "200": {
                        "description": "Listar ingredientes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ingredients"
                ],
                "operationId": "290f4795ec7faed0d1e88d61aeb935ee",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Criado"
                    }
                }
            }
        },
        "/api/ingredients/{id}": {
            "get": {
                "tags": [
                    "Ingredients"
                ],
                "operationId": "0a054220b0a2b30633ee13a9d92f6a97",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhe"
                    }
                }
            },
            "put": {
                "tags": [
                    "Ingredients"
                ],
                "operationId": "b2b705f67b293905971309a7927b3636",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Atualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ingredients"
                ],
                "operationId": "d7a227e0854bd498dfe1c1b9155b4bd0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removido"
                    }
                }
            }
        },
        "/api/ingredient-prices": {
            "get": {
                "tags": [
                    "Suppliers"
                ],
                "operationId": "6103ca990aabff29d05dce2d5d62b057",
                "responses": {
                    "200": {
                        "description": "Lista"
                    }
                }
            },
            "post": {
                "tags": [
                    "Suppliers"
                ],
                "operationId": "02ce75f4c501f73e5ad32ddccd35d562",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Criado"
                    }
                }
            }
        },
        "/api/ingredient-prices/{id}": {
            "get": {
                "tags": [
                    "Suppliers"
                ],
                "operationId": "01d211141ed72193369d58fbd4d5fb89",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhe"
                    }
                }
            },
            "put": {
                "tags": [
                    "Suppliers"
                ],
                "operationId": "ab2c4ce7e6de1d5b06f76a99faec822e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Atualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Suppliers"
                ],
                "operationId": "8ecf756ca92abc623666554dbf159a49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removido"
                    }
                }
            }
        },
...
    },
...
}